requirement.txt

deepface==0.0.93
numpy==2.1.3
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
streamlit==1.44.1
textblob==0.19.0
